update pipeline query fields
